fix float tests and added JsonRpcHandlerBase
