+++ Json-Rpc/JsonRpcService.cs
@@ -1,18 +1,19 @@
 ﻿namespace AustinHarris.JsonRpc
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reflection;
-    using AustinHarris.JsonRpc;
-
+    /// <summary>
+    ///     For routing use SessionId
+    /// </summary>
     public abstract class JsonRpcService
     {
         protected JsonRpcService()
         {
-             ServiceBinder.BindService(Handler.DefaultSessionId(), this);
+            ServiceBinder.BindService(Handler.DefaultSessionId(), this);
         }
 
+        /// <summary>
+        ///     Routing by SessionId
+        /// </summary>
+        /// <param name="sessionID"></param>
         protected JsonRpcService(string sessionID)
         {
             ServiceBinder.BindService(sessionID, this);
